requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
langchain
openai
faiss-cpu
tiktoken
loguru
prometheus-client
